Loading and prevent rerender
- show loading icon and overlay when saving/submitting
- prevent sidebar rerender by lifting the sidebar component up
