+++ src/lib/posts.ts
@@ -3,7 +3,7 @@ import {
   collection,
   deleteDoc,
   doc,
-  updateDoc,
+  updateDoc
 } from "firebase/firestore";
 import { getDownloadURL, ref, uploadBytes } from "firebase/storage";
 import { db, storage } from "./firebase";
