+++ src/styles/globals.css
@@ -9,13 +9,6 @@ form {
   flex-direction: column;
 }
 
-.MuiBox-root .MuiBox-root {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin: 1rem;
-}
-
 .MuiBox-root .MuiTypography-body2 {
   max-width: 80%;
 }
@@ -25,9 +18,4 @@ form .MuiFormControl-root {
   align-items: center;
   justify-content: space-between;
   margin: 0.5rem;
-}
-
-form .MuiInput-root {
-  width: 50%;
-  margin-left: 18rem;
 }
